requirements.txt
# requirements.txt for Segment Anything Optimized (SAM+)
# Supports: SOTA reproduction, optimization, deployment, and optional PyPI packaging

# ===============================
# 🔧 Core Dependencies (Required)
# ===============================
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.23
scipy>=1.10
Pillow>=9.5
matplotlib>=3.7
opencv-python-headless>=4.7.0  # Headless for server compatibility

# ===============================
# ⚡ Training Framework
# ===============================
pytorch-lightning>=2.2.1
lightning-utilities>=0.10
torchmetrics>=1.3.2

# ===============================
# 📊 Logging & Sweeps
# ===============================
wandb>=0.16.3

# ===============================
# 🌐 Inference API (FastAPI)
# ===============================
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.7
pydantic>=2.7

# ===============================
# 📓 Jupyter + Report Export
# ===============================
jupyter>=1.0
nbconvert>=7.10

# ===============================
# 🤗 Hugging Face Integration
# ===============================
huggingface_hub>=0.22

# ===============================
# 🔄 Quantization + ONNX Export
# ===============================
onnx>=1.15
onnxruntime>=1.17

# ===============================
# 🛠 Utility Libraries
# ===============================
tqdm>=4.66
requests>=2.31
pyyaml>=6.0
rich>=13.7
GitPython>=3.1
jsonschema>=4.21

# ===============================
# 🔔 Optional: Slack Notifications
# ===============================
# [notifications]
slack_sdk>=3.27

# ===============================
# 📈 Optional: ML Evaluation
# ===============================
# [analysis]
scikit-learn>=1.4

# ===============================
# 🖼️ Optional: Local UI Demos
# ===============================
# [demo]
gradio>=4.25

# ===============================
# 📦 Packaging, Linting, CI
# ===============================
packaging>=24.0
setuptools>=70.0
wheel>=0.43
pytest>=8.1
autopep8>=2.1
pre-commit>=3.7

# ===============================
# ⚙️ CUDA Compatibility (Optional)
# ===============================
# For CUDA-enabled environments (uncomment if using conda-based install)
# cudatoolkit=11.8
